requirements.txt
# 核心依赖
transformers>=4.30.0
torch>=2.0.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
modelscope>=1.9.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# 可选依赖
tqdm>=4.65.0
